[TRL-317] fix: 인증 객체로 토큰 생성하는 기능 제거
