+++ src/main/java/com/cosain/trilo/auth/infra/TokenProvider.java
@@ -1,10 +1,7 @@
 package com.cosain.trilo.auth.infra;
 
-import org.springframework.security.core.Authentication;
 
 public interface TokenProvider {
-    String createAccessToken(final Authentication authentication);
     String createAccessTokenById(final Long id);
-    String createRefreshToken(final Authentication authentication);
     String createRefreshTokenById(final Long id);
 }
