+++ src/test/java/com/cosain/trilo/unit/auth/infra/jwt/JwtTokenAnalyzerTest.java
@@ -1,25 +1,27 @@
-package com.cosain.trilo.unit.auth.application;
+package com.cosain.trilo.unit.auth.infra.jwt;
 
-import com.cosain.trilo.auth.infra.jwt.JwtTokenAnalyzer;
-import com.cosain.trilo.auth.infra.jwt.JwtTokenProvider;
 import com.cosain.trilo.auth.infra.TokenAnalyzer;
 import com.cosain.trilo.auth.infra.TokenProvider;
-import com.cosain.trilo.support.auth.AuthHelper;
+import com.cosain.trilo.auth.infra.jwt.JwtTokenAnalyzer;
+import com.cosain.trilo.auth.infra.jwt.JwtTokenProvider;
 import com.cosain.trilo.user.domain.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.security.core.Authentication;
 
 import static com.cosain.trilo.fixture.UserFixture.KAKAO_MEMBER;
 
 class JwtTokenAnalyzerTest {
 
-    private static final String SECRET_KEY = "K".repeat(32);
+    private final String SECRET_KEY = "K".repeat(32);
+    private final int ACCESS_TOKEN_EXPIRY = 1000;
+    private final int REFRESH_TOKEN_EXPIRY = 10000;
     private TokenAnalyzer tokenAnalyzer;
+    private TokenProvider tokenProvider;
 
     @BeforeEach
     void setUp(){
+        tokenProvider =  new JwtTokenProvider(ACCESS_TOKEN_EXPIRY, REFRESH_TOKEN_EXPIRY, SECRET_KEY);
         tokenAnalyzer = new JwtTokenAnalyzer(SECRET_KEY);
     }
 
@@ -39,7 +41,7 @@ class JwtTokenAnalyzerTest {
     void 토큰에서_ID_추출하기(){
         // given
         User user = KAKAO_MEMBER.create();
-        String accessToken = createAccessToken(user, 100000L, 1000000L, SECRET_KEY);
+        String accessToken = tokenProvider.createAccessTokenById(user.getId());
 
         // when
         Long userId = tokenAnalyzer.getUserIdFromToken(accessToken);
@@ -52,7 +54,7 @@ class JwtTokenAnalyzerTest {
     void 토큰_유효성_검사_성공시_TRUE를_반환한다(){
         // given
         User user = KAKAO_MEMBER.create();
-        String accessToken = createAccessToken(user, 100000L, 1000000L, SECRET_KEY);
+        String accessToken = tokenProvider.createAccessTokenById(user.getId());
 
         // when
         boolean result = tokenAnalyzer.validateToken(accessToken);
@@ -64,8 +66,9 @@ class JwtTokenAnalyzerTest {
     @Test
     void 토큰_유효성_검사시_secretKey가_다르면_FALSE를_반환한다(){
         // given
+        tokenProvider = new JwtTokenProvider(ACCESS_TOKEN_EXPIRY, REFRESH_TOKEN_EXPIRY, "S".repeat(32));
         User user = KAKAO_MEMBER.create();
-        String accessToken = createAccessToken(user,100000L, 100000L, "S".repeat(32));
+        String accessToken = tokenProvider.createAccessTokenById(user.getId());
 
         // when
         boolean result = tokenAnalyzer.validateToken(accessToken);
@@ -78,7 +81,8 @@ class JwtTokenAnalyzerTest {
     void 만료된_토큰_유효성_검사시_토큰_만료기간이_지났다면_FALSE를_반환한다(){
         // given
         User user = KAKAO_MEMBER.create();
-        String accessToken = createAccessToken(user, 0L, 0L, SECRET_KEY);
+        tokenProvider = new JwtTokenProvider(0, 0, SECRET_KEY);
+        String accessToken = tokenProvider.createAccessTokenById(user.getId());
 
         // when
         boolean result = tokenAnalyzer.validateToken(accessToken);
@@ -91,7 +95,7 @@ class JwtTokenAnalyzerTest {
     void 토큰에서_유효기간_추출하기(){
         // given
         User user = KAKAO_MEMBER.create();
-        String accessToken = createAccessToken(user, 100000L, 1000000L, SECRET_KEY);
+        String accessToken = tokenProvider.createAccessTokenById(user.getId());
 
         // when
         Long tokenExpiryFrom = tokenAnalyzer.getTokenRemainExpiryFrom(accessToken);
@@ -100,10 +104,5 @@ class JwtTokenAnalyzerTest {
         Assertions.assertThat(tokenExpiryFrom).isNotNull();
     }
 
-    private String createAccessToken(User user, Long accessTokenExpiry, Long refreshTokenExpiry, String secretKey){
-        TokenProvider tokenProvider = new JwtTokenProvider(accessTokenExpiry, refreshTokenExpiry, secretKey);
-        Authentication authentication = AuthHelper.createAuthentication(user);
-        return tokenProvider.createAccessToken(authentication);
-    }
 
 }
